Removed code relative to Scraper; promoted remaining sources and removed useless api package.
