--- src/main/resources/bands.json
@@ -1,310 +1,0 @@
-[
-  "1349",
-  "7 Weeks",
-  "Abbath",
-  "Aborted",
-  "AC/DC",
-  "Accept",
-  "Aerosmith",
-  "Airbourne",
-  "Alice in Chains",
-  "Alice Cooper",
-  "Amenra",
-  "American Head Charge",
-  "Amon Amarth",
-  "Andréas & Nicolas",
-  "Animals as Leaders",
-  "Annihilator",
-  "Arch Enemy",
-  "ASG",
-  "At the Gates",
-  "Audioslave",
-  "Avenged Sevenfold",
-  "Baroness",
-  "Bathory",
-  "Batushka",
-  "Behemoth",
-  "Benighted",
-  "Biffy Clyro",
-  "Biohazard",
-  "Black Sabbath",
-  "Black Label Society",
-  "Blind Guardian",
-  "Blink-182",
-  "Blue Öyster Cult",
-  "Burzum",
-  "Candlemass",
-  "Cannibal Corpse",
-  "Carcass",
-  "Carpenter Brut",
-  "Catamenia",
-  "Cavalera Conspiracy",
-  "Celtic Frost",
-  "Children of Bodom",
-  "Clawfinger",
-  "Clutch",
-  "Coal Chamber",
-  "Converge",
-  "Corrosion of Conformity",
-  "Cradle of Filth",
-  "Crowbar",
-  "Cult of Luna",
-  "Cynic",
-  "Danko Jones",
-  "Dark Tranquility",
-  "Darkane",
-  "Darkspace",
-  "Dead Can Dance",
-  "Decapitated",
-  "Deep Purple",
-  "Def Leppard",
-  "Deftones",
-  "Devil Driver",
-  "Devin Townsend",
-  "The Dillinger Escape Plan",
-  "Dio",
-  "Dire Straits",
-  "Disturbed",
-  "Down",
-  "Dream Theater",
-  "Dying Fetus",
-  "Eagles of Death Metal",
-  "Earth",
-  "Edguy",
-  "Electric Wizard",
-  "Emigrate",
-  "Emperor",
-  "Enslaved",
-  "Entombed",
-  "Epica",
-  "Exodus",
-  "Faith No More",
-  "Fantômas",
-  "Fear Factory",
-  "Finntroll",
-  "Foo Fighters",
-  "Freak Kitchen",
-  "Fu Manchu",
-  "Gamma Ray",
-  "Garbage",
-  "Genesis",
-  "Ghost",
-  "Godsmack",
-  "Gojira",
-  "Gorgoroth",
-  "Grand Magus",
-  "Graveyard",
-  "Guns N' Roses",
-  "Hacride",
-  "Hangman's Chair",
-  "The Haunted",
-  "Hawkwind",
-  "Helmet",
-  "Hermano",
-  "High on Fire",
-  "HIM",
-  "Iced Earth",
-  "Ihsahn",
-  "Immortal",
-  "Impaled Nazarene",
-  "In Flames",
-  "Incubus",
-  "Iron Maiden",
-  "Isis",
-  "Judas Priest",
-  "Kadavar",
-  "Karma To Burn",
-  "Karnivool",
-  "Kataklysm",
-  "Katatonia",
-  "Killswitch Engage",
-  "King Crimson",
-  "King Diamond",
-  "Kiss",
-  "Kittie",
-  "Klone",
-  "KMFDM",
-  "Korn",
-  "Kvelertak",
-  "Kylesa",
-  "Kyuss",
-  "Lacuna Coil",
-  "Led Zeppelin",
-  "Leprous",
-  "Limp Bizkit",
-  "Linkin Park",
-  "Lofofora",
-  "Machine Head",
-  "Magma",
-  "Manilla Road",
-  "Manowar",
-  "Marduk",
-  "Marillion",
-  "Marilyn Manson",
-  "The Mars Volta",
-  "Mass Hysteria",
-  "Mastodon",
-  "Mayhem",
-  "Megadeth",
-  "Melvins",
-  "Mercyful Fate",
-  "Meshuggah",
-  "Metallica",
-  "Ministry",
-  "Misfits",
-  "Monkey3",
-  "Monolord",
-  "Monster Magnet",
-  "Moonspell",
-  "Motörhead",
-  "Mötley Crüe",
-  "Mr. Bungle",
-  "Mudvayne",
-  "Mudweiser",
-  "Muderdolls",
-  "Muse",
-  "Mushroomhead",
-  "Mutoid Man",
-  "My Dying Bride",
-  "Naplam Death",
-  "Nasum",
-  "Ne Obliviscaris",
-  "Nebula",
-  "Neurosis",
-  "Nevermore",
-  "Nightwish",
-  "Nile",
-  "Nine Inch Nails",
-  "Nirvana",
-  "Nostromo",
-  "Obituary",
-  "The Ocean",
-  "Om",
-  "Opeth",
-  "Orange Goblin",
-  "Oranssi Pazuzu",
-  "Orden Ogan",
-  "Orgy",
-  "OSI",
-  "Otep",
-  "Pain of Salvation",
-  "Pallbearer",
-  "Pantera",
-  "Paradise Lost",
-  "Papa Roach",
-  "Pearl Jam",
-  "Pentagram",
-  "Pelican",
-  "A Perfect Circle",
-  "Periphery",
-  "Pink Floyd",
-  "Planet X",
-  "The Police",
-  "Porcupine Tree",
-  "Powerwolf",
-  "Primal Fear",
-  "Primus",
-  "Prong",
-  "Pryapisme",
-  "Punish Yourself",
-  "Puscifer",
-  "Queen",
-  "Queens of The Stone Age",
-  "Queensrÿche",
-  "Radiohead",
-  "Rage",
-  "Rage Against The Machine",
-  "Rainbow",
-  "Ramesses",
-  "Rammstein",
-  "Red Fang",
-  "Red Hot Chili Peppers",
-  "Refused",
-  "Rhapsody",
-  "Rose Tatoo",
-  "Rotting Christ",
-  "Rush",
-  "Russian Circles",
-  "Sabaton",
-  "Saint Vitus",
-  "Samael",
-  "Satyricon",
-  "Savatage",
-  "Saxon",
-  "Scorpions",
-  "Septicflesh",
-  "Sepultura",
-  "Shinedown",
-  "Shining",
-  "Shrinebuilder",
-  "Silverchair",
-  "Sinch",
-  "Slayer",
-  "Sleep",
-  "Sleeppers",
-  "Slipknot",
-  "Smashing Pumpkins",
-  "Snot",
-  "Soilwork",
-  "Sonata Arctica",
-  "Sons of Apollo",
-  "Soulfly",
-  "Soundgarden",
-  "Sólstafir",
-  "Spiritual Beggars",
-  "Spock's Beard",
-  "Static-X",
-  "Steel Panther",
-  "Stone Sour",
-  "Suffocation",
-  "Suicidal Tendencies",
-  "Sunn O)))",
-  "Supertramp",
-  "Symphony X",
-  "System of a Down",
-  "Tenacious D",
-  "Tenhi",
-  "Tesseract",
-  "Textures",
-  "Them Crooked Vultures",
-  "Therion",
-  "Thin Lizzy",
-  "Thirty Seconds to Mars",
-  "Thou Art Lord",
-  "Tool",
-  "Trepalium",
-  "Treponem Pal",
-  "Triggerfinger",
-  "Triptykon",
-  "Trivium",
-  "Truckfighters",
-  "Trust",
-  "Tsjuder",
-  "Turbonegro",
-  "Twisted Sister",
-  "Type O Negative",
-  "Ulcerate",
-  "Ultra Vomit",
-  "Ulver",
-  "Uncle Acid",
-  "Unida",
-  "Unsane",
-  "Uriah Heep",
-  "Van Halen",
-  "The Velvet Underground",
-  "Venom",
-  "Vista Chino",
-  "Voivod",
-  "W.A.S.P.",
-  "White Zombie",
-  "Whitesnake",
-  "Windhand",
-  "Witchcraft",
-  "Within Temptation",
-  "Wolfmother",
-  "X Japan",
-  "Yes",
-  "YOB",
-  "Young Gods",
-  "Zu"
-]
