--- src/main/scala/com/music/api/model/entities/types/Album.scala
@@ -1,3 +1,0 @@
-package com.music.api.model.entities.types
-
-case class Album(uuid: Option[String], title: String, releaseDate: Option[String])
