--- src/main/scala/com/music/api/model/entities/types/Band.scala
@@ -1,8 +1,0 @@
-package com.music.api.model.entities.types
-
-case class Band(uuid: Option[String],
-                name: String,
-                aka: Option[String],
-                country: Option[String],
-                formed: Option[String],
-                disbanded: Option[String])
