--- src/main/scala/com/music/api/model/entities/types/Person.scala
@@ -1,8 +1,0 @@
-package com.music.api.model.entities.types
-
-case class Person(uuid: Option[String],
-                  full_name: String,
-                  first_name: Option[String],
-                  last_name: Option[String],
-                  aka: Option[String],
-                  born: Option[String])
