--- src/main/scala/com/music/api/model/relationships/types/PlayedIn.scala
@@ -1,3 +1,0 @@
-package com.music.api.model.relationships.types
-
-case class PlayedIn(instruments: Option[String], start: Option[String], end: Option[String])
