+++ src/main/scala/com/music/model/entities/types/Album.scala
@@ -1,0 +1,3 @@
+package com.music.model.entities.types
+
+case class Album(uuid: Option[String], title: String, releaseDate: Option[String])
