+++ src/main/scala/com/music/model/relationships/types/PlayedIn.scala
@@ -1,0 +1,3 @@
+package com.music.model.relationships.types
+
+case class PlayedIn(instruments: Option[String], start: Option[String], end: Option[String])
