--- src/main/scala/com/music/scraper/model/BandPage.scala
@@ -1,9 +1,0 @@
-package com.music.scraper.model
-
-case class BandPage(name: String,
-                    formed: Formed,
-                    disbanded: Disbanded,
-                    members: List[Member] = List.empty,
-                    albums: List[Release] = List.empty)
-
-case class Release(title: String, date: String)
