--- src/main/scala/com/music/scraper/model/Disbanded.scala
@@ -1,11 +1,0 @@
-package com.music.scraper.model
-
-case class Disbanded(date: Option[Int])
-
-object Disbanded extends YearExtractor {
-  def apply(raw: String): Disbanded = {
-    val year = extractYear(raw)
-
-    Disbanded(date = year)
-  }
-}
