--- src/main/scala/com/music/scraper/model/Formed.scala
@@ -1,11 +1,0 @@
-package com.music.scraper.model
-
-case class Formed(date: Option[Int], location: Option[Location] = None)
-
-object Formed extends YearExtractor {
-  def apply(raw: String): Formed = {
-    val year = extractYear(raw)
-
-    Formed(date = year)
-  }
-}
