--- src/main/scala/com/music/scraper/model/Location.scala
@@ -1,3 +1,0 @@
-package com.music.scraper.model
-
-case class Location(country: String, state: Option[String], city: Option[String])
