--- src/main/scala/com/music/scraper/model/YearExtractor.scala
@@ -1,9 +1,0 @@
-package com.music.scraper.model
-
-abstract class YearExtractor {
-  def extractYear(raw: String): Option[Int] = {
-    val rgxYear = """(\d{4})""".r
-
-    rgxYear.findFirstIn(raw).map(_.toInt)
-  }
-}
