+++ src/test/scala/com/music/model/entities/queries/PersonQueriesTest.scala
@@ -1,0 +1,12 @@
+package com.music.model.entities.queries
+
+import akka.http.scaladsl.testkit.Specs2RouteTest
+import org.specs2.Specification
+
+class PersonQueriesTest extends Specification with Specs2RouteTest {
+
+  def is = s2"""$findExisting"""
+
+  private def findExisting = ???
+
+}
